Move multicastNow to new file
