+++ run/src/run.ts
@@ -1,5 +1,6 @@
-import { merge, makeReplaySubject, Producer } from '@cycle/callbags';
+import { merge, makeReplaySubject } from '@cycle/callbags';
 import { Plugin, Main, MasterWrapper, Subscription } from './types';
+import { multicastNow } from './multicastNow';
 
 let currentId = 0;
 
@@ -8,54 +9,6 @@ function cuid(): number {
     currentId = 0;
   }
   return currentId;
-}
-
-export function multicastNow<T>(source: Producer<T>): Producer<T> {
-  let sinks: any[] = [];
-  let last: any = undefined;
-  let hasLast = false;
-
-  let talkback: any;
-
-  source(0, (t, d) => {
-    if (t === 0) {
-      talkback = d;
-    } else {
-      if (t === 1) {
-        Promise.resolve().then(() => {
-          hasLast = false;
-          last = void 0;
-        });
-
-        last = d;
-        hasLast = true;
-      }
-
-      let hasDeleted = false;
-      for (const sink of sinks) {
-        if (sink) sink(t, d);
-        else hasDeleted = true;
-      }
-
-      if (hasDeleted) {
-        sinks = sinks.filter(Boolean);
-      }
-    }
-  });
-
-  return (_, sink) => {
-    sinks.push(sink);
-    sink(0, () => {
-      sinks[sinks.indexOf(sink)] = void 0;
-      if (sinks.every(x => x === undefined)) {
-        talkback(2);
-        sinks = [];
-      }
-    });
-    if (hasLast) {
-      sink(1, last);
-    }
-  };
 }
 
 export function run(
